MachineLearningProject/requirements.txt
scikit-learn==1.5.0
scikit-image~=0.24.0
xgboost==2.1.1
imbalanced-learn==0.13.0
PyYAML==6.0.2
matplotlib==3.9.2
seaborn==0.13.2
numpy~=2.0.2
Pillow==10.4.0


imblearn~=0.0
albumentations~=2.0.8